refactor: dissolve views/source/ — sourcePicker + loadingOverlay → views/widgets/
Both are reusable DOM components; "source" was a feature folder
breaking the codebase's type-based grouping rule. Landing them in
views/widgets/ first; Task 6 renames widgets/ → components/.
