+++ app/src/main.ts
@@ -22,8 +22,8 @@ import { applyHljsTheme } from './utils/syntaxTheme.js';
 import { buildIconAtlas } from './scene/components/buildings/iconAtlas.js';
 import { setIconAtlas } from './scene/components/buildings/buildings.js';
 import { setCellIconAtlas } from './scene/components/buildings/buildingsCell.js';
-import { createSourcePicker, type SourcePayload } from './views/source/sourcePicker.js';
-import { createLoadingOverlay } from './views/source/loadingOverlay.js';
+import { createSourcePicker, type SourcePayload } from './views/widgets/sourcePicker.js';
+import { createLoadingOverlay } from './views/widgets/loadingOverlay.js';
 import { streamManifest } from './api/manifest.js';
 import { pushRecent } from './utils/sourceRecents.js';
 import { startRenderLoop, _applyDisplayLabel } from './scene/renderLoop.js';
