+++ app/src/views/widgets/loadingOverlay.ts
@@ -1,4 +1,4 @@
-// loadingOverlay.ts — Centered spinner + stepped progress indicator shown
+// views/widgets/loadingOverlay.ts — Centered spinner + stepped progress indicator shown
 // whenever a manifest is being fetched or applied. Reused for direct-boot
 // loads and modal submits so the user sees the same UI regardless of entry
 // point.
